GitHub actions for updating DrawnToDigital/homebrew-pubman (#52)
Relates to https://github.com/DrawnToDigital/pubman/issues/44

This pull request updates the release workflow to automate the process
of updating the Homebrew cask
https://github.com/DrawnToDigital/homebrew-pubman. The changes include
calculating SHA256 checksums for release files, checking out the
Homebrew repository, and modifying the cask file with the new version
and checksums.

### Workflow automation for Homebrew cask updates:

* **SHA256 checksum calculation**: Added a step to compute SHA256
checksums for the ARM and Intel macOS `.dmg` files and store them as
outputs for later use.
* **Homebrew repository checkout**: Added a step to clone the
`homebrew-pubman` repository using the `actions/checkout` action.
* **Cask file update**: Added a step to update the `pubman.rb` cask file
with the new version and computed SHA256 checksums, and then commit and
push the changes if the workflow is triggered by a version tag.